+++ boofuzz/fuzz_logger_csv.py
@@ -76,6 +76,9 @@ class FuzzLoggerCsv(ifuzz_logger_backend.IFuzzLoggerBackend):
     def close_test(self):
         pass
 
+    def window_resize(self):
+        pass
+
     def _print_log_msg(self, msg):
         time_stamp = get_time_stamp()
         self._csv_handle.writerow([time_stamp] + msg)